Replacing Future with proper mockers
In a different development, the Future are starting to cause problems.

This is due to the fact that Future have a very specific place in the event_loop system. Creating them will implicitly attach them to event_loops
and expect those event_loops to interact with those futures in a certain way

So it is not a good practice to just instantiate them in the tests

So we are replacing them with AsyncMocks, which does exactly what we were trying to make Futures do: represent async functions.

To do that we need to:

- Upgrade mock to at least 4.0 to use AsyncMock
- Upgrade pytest-mock to at least 1.4 to use mock as its fixture source
- Set mock_use_standalone_module to true to replace unittest.mock with mock lib
- Add a 'spec' source on places where a MagicMock should know that a specific function is an asynchronous one