Attempting python3.8 to see if it works
Python3.8 seems to have some improvements asyncio-wise. This might be a solution to a problem we are having now with celery handling event loops

Basically, we recently implemented `asyncio.run` to replace manual handling of event loops, because that's python suggested way of doing this.

We know this fixes issues where async code from the previous task was running when a new task had already started. So we imagine it would solve at least some timing issues.

But now we have big queues of tasks happening on celery.

We suspect this is because now the system wait for all async code to be terminated before the tasks are declared as finished. And sending cancallation signals to those async codes.

And the cancellation signals are exceptions. And we have a lot of `except Exception` blocks that are possibly swallowing and silently ignoring those exceptions, and not stopping.

On python3.8 taht is not possible because on 3.8 `CancelledError` does not excent from `Exception`, but instead from `BaseException`. So those errors cannot be easily ignored.

Also upgrading `psycopg` to a newer version because the previous one wouldnt run with python3.8

Running the image locally worked.