Using admin-level bot on hook creation
For creating a webhook, we cannot use the public bot, since it does not have permissions to create hooks on any repository it wants.

So we are setting a mroe specific admin-bot-token for those cases.

This will regardless be replaces soon when we develop a more finer grain control of bots per read/write endpoint