Setting worker_max_memory_per_child
Attempting to limit the amount of memory each process uses.

We are having "Out of memory" errors still. There is probably some memory leak somehwere we don't understand.

There are probably a lot of improvements on how we can use memory, but for now we are just going to make each child processed be killed after they use too much memory. This way the machine as a whole doesn't starve and the main process is never killed.

We are also logging the size of each object downloaded to see if they are the culprit (some big files are killing the workers), or if a slow memory leak is being the problem