Trying piptools
piptools seems to be a tool that works on doing what we want to do with `requirements.txt` and `manual_requirements.txt`: it separates upgrading/managing dependencies from keeping reproducible builds.

It does that by having a standard name for our `manual_requirements`: it is called `requirements.in`

`requirements.in` is the file that we are supposed to update manually. Then calling the command `pip-compile requirements.in` regenerates the `requirements.txt` with the proper values