Replacing pip download with pip wheel
It seems like now more and more packages are using different 'build dependencies'. Most recently is is 'multidict'

On those build dependencies, the packages require extra packages to be built. And such extra packages are not downloaded properly with 'pip download'.

- For example. `multidict` kept requiring a specific `pip` version that was not in the `pip-package` folders we had the dependencies downloaded on.
- We had also that problem in the past with the `cryptography` package, but we added `pip download wheel` as a workaround.

The `pip wheel` answer to this problem is also a good answer to the original problem that 'pip download' tries to solve:

'pip wheel' not only downloads the dependencies we will need in the future, but it already generates a 'wheel' (special single python file for using libs) from them.

This command takes care of not only downloading the proper 'build dependencies', but it also already uses them to do the actual build, and does not add them to the runtime image (since they are build-only). So the files generated by this step are alreaty the libs in thei final state ready to be used.

This for example removes the need for the `pip download wheel setuptools' that we had

So very likely, the `pip wheel` command is a good fit to solving that problem