Removing pull.base from squash calculation
We currently use a list of the PR commits to determine if it was squash-merged or not.

Due to a mistake, the base commit was also included at that list. And since the base commit does not belong to the PR anyway, it would wrongly tell some commits were normal-merged other than squash merged