Implementing multiple types of github tokens to use on shared
Fetching alternative bot keys to use on public repositories (specially for github)

For now we are only using one bot per type. To includem ore bots per type we need to decide on a clean way to allow arrays on configuration via env vars.

But this should be enough to get us started