Using particular bot in more situations
After seeing the initial results of the token splitting on Github we noticed:

- Public bots cannot set statuses on public repos (although they can read them)
- The majority of the bot use are reads (code or diff or PRs listing), making it probably use the whole quota by itself.

For the first issue, the solution is to try to use the actual particular bot (bot associated with the repo or its owner) on the status setting

For the second issue, the 'ideal' solution is to simply split the load among more bots. As we said before, this split takes some work because we need a clean way to make lists on env vars.

I do have a nice idea on how to do that splitting on envvars. But first, I'd like to try the following for the reads: use the particular user bot, if that is available.

Although this behavior is not perfect (in case of buggy setting of bots like we had in the past, we could have some Github accounts reading information from public repos that is not related to them. It's not damaging to anyone, since those are not only reads on repos taht are public anyway), this matches exactly what legacy was doing: it used the particular bot on everything possible except for the commenting itself.

So this might be a way of distributing the load: using the public repos actual bots if possible for the read operations instead of automatically using the public bot for reading from everyone.