Added repository coverage and filtering by names
This PR adds the ability to filter which repos are returned from the repository LIST endpoint by name. Using a query param like `?names=codecov-api,codecov.io` we can limit which repositories are returned by the list.

This PR also adds the ability to fetch the coverage data of the latest commit of each repo when listing. This will help complete the link issue which needs coverage data for multiple repos in a single request.

These changes were necessary to complete the work in the linked issue as the list of repositories has to be filtered by name.

Related to https://codecovio.atlassian.net/browse/CE-1873