Using httpx inside Notifiers
We have recently had a lot of Hard timeouts on some `NotifyTask`s. This happens because some notifiers, like the SlackNotifier, use requests.post, which does not allow itself to be canceled by async cancelations, since it is not an async function.

One good idea is to start using httpx. It is meant as an almost-dropin replacement for requests, with async support.

This is also good testing grounds to replace `tornado` http client with `httpx` in the future. `httpx` is not a fully framework that we use only a small part of. And it is done by `encode`, which already produces software we use.