Ingesting Pathmap
Ingesting pathmap so it stops being an external dependency and becomes something we can properly manage.

It looks like that every once in a while we depend on something on pathmap and it becomes a pain to manage it/update it/replace it the way we want to. So the purpose of this change is to put it inside our codebase so we can slowly replace its parts with something we really understand well and can properly manage.