Add a filter to remove repositories with no coverage
For the analytics page, we use the totals field on the repositories to display the coverage values of each repo. It is highly possible that repos don't have coverage data for the time selected or at all.

If we filter those repository out in the client, this can mess with the pagination and the views. To prevent that, we added a filter on the backend to filter those repos out if a query param is passed.