Decreasing max_retries
When we made the change on https://github.com/codecov/worker/pull/377 , we allowed for more retries from heavy customers. As described on https://github.com/codecov/worker/pull/383 , that caused some problems.

But since this was enough of a problem to decide for a rollback, I think that's better to go more conservative instead. So I am reducing the number of retries. This is still better than before since the starting point is 200 seconds. Also halving the 'delay' of 200 s to a 60s in case there is some processing already happening