Accounting for possible repeated lines
So far, it seems that the gcov parser always assumed that each line would appear only once in the report (outside of branch and loop cases), and always would assume it was safe to just store the last occurrenc of the line.

At least according to gs://codecov/v4/raw/2020-06-05/83958652AC5FFFFC5FEFAF895B5F3800/1d9dfe305fdaaa987df95f71c7a066791d2a9d58/df36eac5-d822-4620-93e1-7875cdf065ed.txt that is not true. On file 'rsl/h264/Mp4NaluParser.h' it shows that line 13 is listed multiple times, is not a branch and is supposed to be called coverage.

So we are accounting for the multiple occurrences of the same line by using a list of 'coverages'

Unfortunately I don't understand the gcov format in deep to see if that is indeed what needs to be accounted for. But this seemingly did not break other test cases, and accounted properly for this event