Using Report Models on Processing
This change includes using the report models that were included in the system a while ago on https://github.com/codecov/codecov.io/pull/675

The usage of the models itself is pretty straightforward: whatever information is on the current `commit.report_json` should be also on those new tables.

But this change is a little more complex because of an inversion of order.

Currently, all we do is take the job properties (url, build, job etc) from redis and pass them as arguments to the `UploadProcessorTask`. Then it processes, generates a full-blown report. From that report, it generates all the data that needs to be on `commit.report_json`

We want to invert that order. We want to put the job details on database first, then move on to process the report and update the database with the results.

The main gain from this is that the "upload job" will now stop being something that "goes from redis when created to the database when finished" to start being something that is "is created on the database and is later updated on the database as the processing goes".

 The change might look subtle, but it gives:

- Permanentability: the job will live permanently on the db, along with its statuses and problems. This gives us more ability to show them in the UI or wherever we need to see them.
- Reproducibility: the user will be able to try to re-run their jobs, and support might be able to pick specific uploads to re-run and see the results. Ability to see a job's result will not be limited to the moment it is uploaded anymore

This then becomes the flow:

1) The job, once it hits the worker, needs to be stored on the database as soon as the `UploadTask` schedules it. It will also schedule it with a state of `started`. Then all it needs to do is to send the `id` of the database object to the `UploadProcessorTask` (right now it still sends the other stuff, but this is more a safety step we will remove as we go).
2) The `UploadProcessorTask` will take the upload id, retrieve the object, and do the processing with its information. When that processing for that step is done, it can set the `upload` object state to `completed` or `error` according to whatever happens inside it. In the future, it will also be able to create the errors and attach to this `upload` object.

But by also giving us those abilities, it requires some important changes:

1) We need to backfill for every commit that comes in from the "old way".
2) Similarly, but in a different way, reports that are carryforwarded can't really have their `upload` corresponding objects created before the CFF is calculated, so they still need some form of "backfilling" the uploads after the CFF calculation is done

Both of those steps will be done at the `UploadTask`, which sounds like the better place for the CFF calculation anyway.