Covering case otherwise good parent has no report
There can be a case where a parent has all the right fields but not a report saved on storage. This could lead to problems for the child commit CFFing from it. So we are stopping that and makign the child build a brand new report.
This is ok because we expect those situations to not be the norm