Change the filter for the repository list
The previous filter expected you to pass `names=test1,test2,test3,...`, but the way lists are handled in modern URLs is by passing the parameter multiple times, `names=test1&names=test2&names=test3&...`. A custom filter needed to be created to make sure we followed this standard.

This PR creates a custom filter for repository names that will extract the list from the query parameters and filter with that rather than rely on the comma splitting form django.

Work is related to https://codecovio.atlassian.net/browse/CE-2283