Upgrading vcrpy and httpx
Upgrading vcrpy and httpx. The reson is that the current version of vcrpy (the thing that mocks http calls) did not support mocking httpx back then. That's why we would constantly get

```
Coverage for ThiagoCodecov/example-python no change 0.00% on thiago/f/something is 85.00000% via 11daa27

```

messages on slack, even though the vcrpy lib was supposed to be mocking it.

Upgrading httpx is more because I was not sure if the 0.13 version of httpx was the version that the recent version of vcrpy would mock correctly

This will also make testing go faster, because we will keep from doing those real http calls