Add support for sorting repositories by various fields
We wanted to be able to sort repositories by their coverage data using the repository list in the various pages of the new UI.

This PR adds sorting to the repository endpoints using the fields form the latest commit. It will not impact the API too much and should still have the same performance. The implementation can be improved, but it will work for the needs we have right now.

Work is related to https://codecovio.atlassian.net/browse/CE-2201.