Properly implementing get_authenticated_user
When we ported `get_authenticated_user` from legacy, we ended up bringing something that is only useful in legacy context

It was not a problem, since we didn't really use the function.

But now we will use it. So we need to implement it properly, so we are letting the function has a 'code'