Add coverage change to the repository chart endpoint
We have to do quite a lot of potentially buggy math on the frontend to find the coverage change over time for commits. These computations should be done on the backend instead to reduce the possibility of errors.

This PR adds a simple calculation that finds the coverage change over the grouped commits for the repository chart endpoint. This computation should not impact performance since we keep the exact same loops.

Work is related to https://codecovio.atlassian.net/browse/CE-2203