Improving go performance by replacing list with set
Sit down, because this is a long explanation.

The golang coverage format is very peculiar. Here is a sample of it:

```
path/file.go:18.95,22.51 3 0
path/file.go:28.2,29.61 2 0
path/file.go:37.2,37.15 1 0
path/file.go:22.51,24.32 1 3
path/file.go:24.50,26.3 1 0
path/file.go:29.61,31.3 1 0
path/file.go:31.8,32.24 1 0
path/file.go:32.24,34.4 1 0
path/file.go:41.75,45.45 3 0
```

It basically consists of many lines, each which has the format:

```
name.go:line.column,line.column numberOfStatements count
```

What basically each piece of information is is a certain interval, and what is the coverage for that interval. For example, the fourth line of the above examples tells us that there were 3 hits on the statement group that spans from "line 22/column 51" to "line 24. column 3"

From the coding of the processor, we also need to be prepared to overlaps between intervals (although I am unable to find examples of such overlap in the wild).

For whoever is used to coding interviews/competitions, this looks awfully like a variation of that 'overlapping intervals' problem:

```
Given a set of time intervals in any order, merge all overlapping intervals into one and output the result which should have only mutually exclusive intervals.
```

which has a number of solutions and best-solutions, which mostly involve ordering the intervals, and traversing them in a certain way to find the overlaps.

If you understand those type of problems, you also understand that finding the most optimal solutions is not always needed, because in most real-life situations the more naive solutions is enough. So there is no reason to look into that here.

And you are right. But we still need to look into this, because right now the go processors are taking 650 for some cases (three times the production limit) and taking a lot of memory on the cases where the file is 1.5GB. Of those 650 seconds, only 100 seconds are spent parsing the file, and the remaining of the 550 seconds are spent supposedely doing those calculations.

So we found what the current solution is doing.

1) First, the code splits each interval into line intervals. So for example the (18.95,22.51) inverval with coverage 0 on the example above becomes:

```
(18.95,18.NULL) with coverage 0
(19.0,19.NULL) with coverage 0
(20.0,20.NULL) with coverage 0
(21.0,21.NULL) with coverage 0
(22.0, 22.51) with coverage 0
```

2) Then, all those intervals, which now only span 1 lines, are grouped per line. On the example above, line 24 would have

(24.0, 24.32) with coverage 3 from line `path/file.go:22.51,24.32 1 3`
(24.50, 24.NULL) with coverage 0 from line `path/file.go:24.50,26.3 1 0`

3) So, until now, the only thing the current implementation did was to split the file-level overlapping problem into small, per-line overlapping problems.

4) Then, later in the logic, for each line, the implemention proceeds to do the typical brute-force solution for the overlapping problem: split intervals into discrete sets of elements. Let's say that for line 56, there are the following intervals

(56.0, 56.25) with coverage 3
(56.12, 56.70) with coverage 0
(56.20, 56.23) with coverage 0
(56.41, 56.NULL) with coverage 1

Then the code does

(56.0, 56.25) with coverage 3 becomes (56.0)/3, (56.1)/3, (56.2)/3, (56.3)/3 ... (56.23)/3, (56.24)/3, (24.24)/3
(56.20, 56.23) with coverage 2 becomes (56.20)/2, (56.21)/2, (56.22)/2
(56.12, 56.70) with coverage 0 becomes (56.12)/0, (56.13)/0, (56.14)/0, (56.15)/0 ... (56.68)/0, (56.69)/0, (24.69)/0
(56.41, 56.NULL) with coverage 1 becomes (24.41)/1, (24.42)/1, ... until what the system thinks is the end ot the line (which would be 70 at this moment, but that's beside the point)

Then the code takes all those bits and groups them by column. It becoms

0 -> (56.0)/3
1 -> (56.1)/3
...
11 -> (56.11)/3
12 -> (56.12)/3 - (56.12)/0
...
20 -> (56.20)/3 - (56.20)/2 - (56.20)/0
...

5) Then, for each column, the system calls the functions `merge_all`, which itself calls the function `merge_coverage` once for each bit. Meaning that line 20 above calls `merge_coverage` twice. Columns that have 20 bits call `merge_coverage` 19 times. And `merge_coverage` is not an extremely simple function as it should be at this point of the stack. Given how deep we are (once per file, per line, per column on the line, per coverage on the column), this time adds up very quickly.

6) Then, each line does some calculation on the resulting "non-overlapping intervals" to determine its general coverage, which is added to the ReportFile

As you imagine, the whole "convert this interval into smaller intervals, then into individual elements of the interval" is a costly unnecessary operation. A proper implementation should be able to calculate this without so many operations.

Now, onto the sad part: if you understand code coverage, you understand it is rare for a given to line to have 5 separate statement coverage levels, let alone 20 (which would generate the 20 bits on step 5 above). So none of this my hypothetical examples should be relevant

And you are right. Except for one thing: there are many repeated intervals, which does lead to 30s of intervals per line, which makes all of my above calculations happen, a lot.

So, a solution that does the proper least-complexity calculation is not needed. All we have to do is to account for repeated intervals.

Turning a list of intervals into a set of intervals does the trick. The calculation that took 650 seconds and 14.7 GB now takes 120 seconds and 7.4GB (most of it which is not used by the algorithm itself and is being taken care of on a different task)