Making mix improvements on memory existing PR
1. Fixing the problem that affected some customers

The source of error here is that the users reports do not have `<<<<<< EOF` on a line of its own:
```
...
covered="10"/><counter type="CLASS" missed="0" covered="3"/></package><counter type="INSTRUCTION" missed="14" covered="212"/><counter type="LINE" missed="0" covered="32"/><counter type="COMPLEXITY" missed="2" covered="21"/><counter type="METHOD" missed="2" covered="21"/><counter type="CLASS" missed="0" covered="7"/></report><<<<<< EOF
```
which was not expected.

I am adding the code to deal with that, but I think we should look into what uploader is sending this that way and change it.

Our very lightly defined format implies line breaks on those: https://docs.codecov.io/v4.4.0/reference#upload

2. Reverting the XML parser

The XML parser was changed for a purpose that is no longer true. So to make it safer, we are changing it back

3. Trying to use more performant strip

We are trying to basically replace a .strip() call with a version that can better use the fact that we have a newly built BytesIO that we can modify freely

The only thing we are doing different is to try to benefit from the fact that BytesIO
            is a mutable structure (unlike), so we can rstrip it without having to make full copies
            of it

So we replace a stirp with a lstrip and a rstrip, and then change a rstrip to just truncate the BytesIO instead of making full copies of the content inside

On memory tests, this lead to peak memory from n * file_size to (n - 1) * file_size,
            because rstrip would almost always have to create a copy of the original bytes. It
            wouldn't create a full copy if the result was exactly the same as the original,
            but those files almost always would have a trailing line break